course_finder_app/requirements.txt
Flask==3.1.1
chromadb==1.0.11
llama-index==0.12.32
llama-index-embeddings-huggingface==0.5.4
llama-index-llms-huggingface-api==0.5.0
llama-index-vector-stores-chroma==0.4.1
huggingface-hub==0.32.3
python-dotenv==1.1.0
requests==2.32.3          
pandas==2.2.3             
tqdm==4.67.1            
